Fix and catch failed item validation with tags
This should fix a problem where item bounties using tags (e.g. #minecraft:wool) would throw an error because they were being converted to an Identifier, of which hashtags are not a valid part of an identifier. I've changed it so that item bounties with tag content are now correctly parsed similarly to a tag objective, and additionally all of these checks are caught to avoid a runtime exception.